+++ packages/database/src/test/utils/index.ts
@@ -1,6 +1,6 @@
 export * from './helpers/extended-outbox-event.helper.js';
 export * from './helpers/logger-mock.helper.js';
-export * from './helpers/outbox-writer-mock.helper.js';
+export * from './helpers/outbox-repository-mock.helper.js';
 
 export * from './repositories/outbox-extended-test.repository.js';
 export * from './repositories/outbox-test.repository.js';